refactor : member 디렉터리명 변경 (2)
