+++ src/main/java/com/rudolph/Weevo/chat/dto/summary/OpponentSummary.java
@@ -1,6 +1,6 @@
 package com.rudolph.Weevo.chat.dto.summary;
 
-import com.rudolph.Weevo.memberr.domain.Member;
+import com.rudolph.Weevo.temp.domain.Member;
 import lombok.Builder;
 import lombok.Getter;
