+++ src/main/java/com/rudolph/Weevo/chat/service/ChatService.java
@@ -7,8 +7,8 @@ import com.rudolph.Weevo.chat.domain.enums.ChatType;
 import com.rudolph.Weevo.chat.dto.request.ChatMessage;
 import com.rudolph.Weevo.chat.dto.response.ChatMessageResponseDto;
 import com.rudolph.Weevo.chat.repository.ChatRepository;
-import com.rudolph.Weevo.memberr.domain.Member;
-import com.rudolph.Weevo.memberr.service.MemberService;
+import com.rudolph.Weevo.temp.domain.Member;
+import com.rudolph.Weevo.temp.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
