+++ src/main/java/com/rudolph/Weevo/course/service/CourseService.java
@@ -9,8 +9,8 @@ import com.rudolph.Weevo.course.domain.Course;
 import com.rudolph.Weevo.course.domain.MemberCourse;
 import com.rudolph.Weevo.course.repository.CourseRepository;
 import com.rudolph.Weevo.course.repository.CourseSpecification;
-import com.rudolph.Weevo.memberr.domain.Member;
-import com.rudolph.Weevo.memberr.repository.MemberRepository;
+import com.rudolph.Weevo.temp.domain.Member;
+import com.rudolph.Weevo.temp.repository.MemberRepository;
 import com.rudolph.Weevo.global.common.code.ErrorStatus;
 import com.rudolph.Weevo.global.exception.GeneralException;
 import com.rudolph.Weevo.global.service.S3Service;
