+++ src/main/java/com/rudolph/Weevo/global/handler/OAuthLoginSuccessHandler.java
@@ -6,8 +6,8 @@ import com.rudolph.Weevo.auth.dto.info.GoogleUserInfo;
 import com.rudolph.Weevo.auth.dto.info.KakaoUserInfo;
 import com.rudolph.Weevo.auth.dto.info.OAuth2UserInfo;
 import com.rudolph.Weevo.auth.repository.RefreshTokenRepository;
-import com.rudolph.Weevo.memberr.domain.Member;
-import com.rudolph.Weevo.memberr.repository.MemberRepository;
+import com.rudolph.Weevo.temp.domain.Member;
+import com.rudolph.Weevo.temp.repository.MemberRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
