+++ src/main/java/com/rudolph/Weevo/notification/service/NotificationService.java
@@ -3,9 +3,9 @@ package com.rudolph.Weevo.notification.service;
 import com.rudolph.Weevo.auth.security.CustomUserPrincipal;
 import com.rudolph.Weevo.global.common.code.ErrorStatus;
 import com.rudolph.Weevo.global.exception.GeneralException;
-import com.rudolph.Weevo.memberr.domain.Member;
-import com.rudolph.Weevo.memberr.repository.MemberRepository;
-import com.rudolph.Weevo.memberr.service.MemberService;
+import com.rudolph.Weevo.temp.domain.Member;
+import com.rudolph.Weevo.temp.repository.MemberRepository;
+import com.rudolph.Weevo.temp.service.MemberService;
 import com.rudolph.Weevo.notification.domain.Notification;
 import com.rudolph.Weevo.notification.domain.enums.NotiType;
 import com.rudolph.Weevo.notification.dto.NotificationListResponseDto;
