+++ src/main/java/com/rudolph/Weevo/temp/dto/response/MemberInterestTagDto.java
@@ -1,7 +1,7 @@
-package com.rudolph.Weevo.memberr.dto.response;
+package com.rudolph.Weevo.temp.dto.response;
 
-import com.rudolph.Weevo.memberr.domain.Member;
-import com.rudolph.Weevo.memberr.domain.MemberInterestTag;
+import com.rudolph.Weevo.temp.domain.Member;
+import com.rudolph.Weevo.temp.domain.MemberInterestTag;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
