+++ src/main/java/com/rudolph/Weevo/temp/repository/MemberInterestTagRepository.java
@@ -1,7 +1,7 @@
-package com.rudolph.Weevo.memberr.repository;
+package com.rudolph.Weevo.temp.repository;
 
-import com.rudolph.Weevo.memberr.domain.Member;
-import com.rudolph.Weevo.memberr.domain.MemberInterestTag;
+import com.rudolph.Weevo.temp.domain.Member;
+import com.rudolph.Weevo.temp.domain.MemberInterestTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
