+++ src/main/java/com/rudolph/Weevo/temp/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.rudolph.Weevo.memberr.repository;
+package com.rudolph.Weevo.temp.repository;
 
-import com.rudolph.Weevo.memberr.domain.Member;
+import com.rudolph.Weevo.temp.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
