+++ src/main/java/com/rudolph/Weevo/temp/repository/MemberTagRepository.java
@@ -1,4 +1,4 @@
-package com.rudolph.Weevo.memberr.repository;
+package com.rudolph.Weevo.temp.repository;
 
 import com.rudolph.Weevo.tag.domain.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
