+++ src/main/java/com/rudolph/Weevo/temp/service/MemberService.java
@@ -1,11 +1,11 @@
-package com.rudolph.Weevo.memberr.service;
+package com.rudolph.Weevo.temp.service;
 
 import com.rudolph.Weevo.auth.security.CustomUserPrincipal;
-import com.rudolph.Weevo.memberr.domain.Member;
-import com.rudolph.Weevo.memberr.dto.request.InfoRequest;
-import com.rudolph.Weevo.memberr.dto.response.MemberDetailResponse;
-import com.rudolph.Weevo.memberr.dto.response.MemberListResponse;
-import com.rudolph.Weevo.memberr.repository.MemberRepository;
+import com.rudolph.Weevo.temp.domain.Member;
+import com.rudolph.Weevo.temp.dto.request.InfoRequest;
+import com.rudolph.Weevo.temp.dto.response.MemberDetailResponse;
+import com.rudolph.Weevo.temp.dto.response.MemberListResponse;
+import com.rudolph.Weevo.temp.repository.MemberRepository;
 import com.rudolph.Weevo.global.common.code.ErrorStatus;
 import com.rudolph.Weevo.global.exception.GeneralException;
 import com.rudolph.Weevo.tag.domain.Tag;
@@ -14,13 +14,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import com.rudolph.Weevo.memberr.domain.MemberInterestTag;
-import com.rudolph.Weevo.memberr.dto.request.FixProfileRequestDto;
-import com.rudolph.Weevo.memberr.dto.request.UpdateInterestTagRequestDto;
-import com.rudolph.Weevo.memberr.dto.response.MemberInterestTagDto;
-import com.rudolph.Weevo.memberr.dto.response.UserProfileDto;
-import com.rudolph.Weevo.memberr.repository.MemberInterestTagRepository;
-import com.rudolph.Weevo.memberr.repository.MemberTagRepository;
+import com.rudolph.Weevo.temp.domain.MemberInterestTag;
+import com.rudolph.Weevo.temp.dto.request.FixProfileRequestDto;
+import com.rudolph.Weevo.temp.dto.request.UpdateInterestTagRequestDto;
+import com.rudolph.Weevo.temp.dto.response.MemberInterestTagDto;
+import com.rudolph.Weevo.temp.dto.response.UserProfileDto;
+import com.rudolph.Weevo.temp.repository.MemberInterestTagRepository;
+import com.rudolph.Weevo.temp.repository.MemberTagRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.reactive.function.client.WebClient;
